pyproject.toml
[tool.poetry]
name = "pelias.adapter"
version = "0.1.0"
description = "Open Transit Tools - Web API / Controller"
authors = ["Open Transit Tools <info@opentransittools.org>"]
readme = "README.md"
license = "Mozilla-derived (http://opentransittools.com)"
keywords = ["ott", "otp", "services", "transit"]
homepage = "http://opentransittools.com"

classifiers = [
    "Programming Language :: Python",
    "Framework :: Pyramid",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Internet :: WWW/HTTP :: WSGI :: Application"
]

[[tool.poetry.packages]]
include = "com"
from = "."

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.119.0"
uvicorn = "^0.29.0"
sqlalchemy = "^2.0.0"
oracledb = "^1.3.0"
python-dotenv = "^1.0.0"
ott-utils = { git = "https://github.com/OpenTransitTools/utils.git", branch = "master" }
setuptools = "^80.9.0"
pydantic-settings = "^2.11.0"
httpx = "^0.28.1"
gunicorn = "^23.0.0"


[tool.poetry.group.dev.dependencies]
black = "^25.9.0"
flake8 = "^7.3.0"
pytest = "^8.4.2"
coverage = "^7.11.0"
webtest = "^3.0.7"
autoflake = "^2.3.1"
autopep8 = "^2.3.2"
pytest-asyncio = "^1.2.0"
asgi-lifespan = "^2.1.0"

[tool.black]
line-length = 88
target-version = ['py310']
include = '\.pyi?$'


[build-system]
requires = ["poetry-core>=1.7.0"]
build-backend = "poetry.core.masonry.api"

